requirements.txt
pytesseract
pdfminer.six
pandas
openpyxl
